Fix Users#Show page
Change routing so that the user is now directed to the user show page instead of going to the group show page.  This can/might change dpending on UX conversation.  Also, fix how we find the user in the users_controller.  Searching for the user based on current_user.id.