Footer Problems
Buttons that appear floating on bottom when you click on the '+' positions can't be changed.  Tried to create a new class for the buttons, tried to fix the buttons.  Can only move all the buttons.  I think the reason is because they are connected to the parent '+' and they can only be moved together.